+++ ISSUE_TEMPLATE.md
@@ -2,7 +2,7 @@
 
 Replace the values below with your own:
 
-- **Bot Version:** v0.5.2
+- **Bot Version:** v0.7.0
 - **Node Version:** v8.9 (Bot require >= 7.6)
 - **Operating System:** Ubuntu 14.04
 - **Browser:** Google Chrome 64
