+++ TODO.md
@@ -1,5 +1,5 @@
 # Roadmap
 See full roadmap (open task, todo and bugs) in [project page](https://github.com/social-manager-tools/instagram-bot.js/projects?query=is%3Aopen+sort%3Aname-asc).
-* ### [v0.7.0](https://github.com/social-manager-tools/instagram-bot.js/projects/1)
-* ### [v0.8.0](https://github.com/social-manager-tools/instagram-bot.js/projects/2)
-* ### [v0.9.0](https://github.com/social-manager-tools/instagram-bot.js/projects/3)
+* ### [v0.7.X](https://github.com/social-manager-tools/instagram-bot.js/projects/1)
+* ### [v0.8.X](https://github.com/social-manager-tools/instagram-bot.js/projects/2)
+* ### [v0.9.X](https://github.com/social-manager-tools/instagram-bot.js/projects/3)
