+++ config.js.tpl
@@ -1,6 +1,6 @@
 module.exports = {
     // InstagramBot.js Configs
-    "debug": true,
+    "debug": false,
     "login": true,
 
     // Instagram Account
@@ -13,9 +13,9 @@ module.exports = {
     // [WORK] likemode_classic          - select random hashtag from config list and like 1 random photo (of last 20) | 400-600 like/day.
     // [WORK] likemode_realistic        - select random hashtag from config list, like fast 10-12 photo and sleep 15-20min. Sleep at night
     // [WORK] likemode_competitor_users - select account, select random followers, like 10-12 photo and sleep 15-20min.
-    // [TEST] likemode_superlike        - select random hashtag from config list and like 3 random photo of same user | 400-600 like/day.
-    // [TEST] fdfmode_classic           - follow user from random hashtag and defollow after 1h | 300 follow-defollow/day.
-    // [TEST] comment_mode              - leave a comment under the photo via hashtag.
+    // [WORK] comment_mode              - leave a comment under the photo via hashtag.
+    // [TODO] likemode_superlike        - select random hashtag from config list and like 3 random photo of same user | 400-600 like/day.
+    // [TODO] fdfmode_classic           - follow user from random hashtag and defollow after 1h | 300 follow-defollow/day.
     // [TODO] fdfmode_defollowall       - defollow all your following (not defollow users in whitelist) | 30 defollow/hour.
     "bot_mode": "likemode_classic",
     "bot_likeday_min": 400,
@@ -43,6 +43,7 @@ module.exports = {
     // LOG
     "log": {
         "drivers": ["console"], // slack or console
+        "screenshot": false,  // disable or enable screenshot in logs folder
         "channels": {
             "console": "",
             "slack": {
