--- modules/comment_mode_classic/comment_sources/README.md
@@ -1,7 +1,0 @@
-### Here is types source comments
-
-#### Interface required method:
-- get_random_comment : return string type
-
-Currently implemented the following types:
-* [✓] Array - the list of comments is specified in config file.
