--- modules/logger/channels/README.md
@@ -1,5 +1,0 @@
-## Channels interface
-
-#### Channels required method:
-
-- method log is required : params type, func and message.
