+++ modules/mode/likemode_classic.js
@@ -1,55 +1,55 @@
-const LOG_NAME = "like_realistic";
-
-const Manager_state = require("../modules/base/state").Manager_state;
-const STATE = require("../modules/base/state").STATE;
-const STATE_EVENTS = require("../modules/base/state").EVENTS;
-
-// log
-const Log = require("./logger/Log");
-
 /**
- * MODE: likemode_realistic
+ * MODE: likemode_classic
  * =====================
- * Select random hashtag from config list, like fast 10-12 photo and sleep 15-20min. Sleep at night | 400-600 like/day.
+ * Select random hashtag from config list and like 1 random photo (of last 20) | 400-600 like/day.
  *
  * @author:     Patryk Rzucidlo [@ptkdev] <support@ptkdev.io> (https://ptkdev.it)
  * @license:    This code and contributions have 'GNU General Public License v3'
- * @version:    0.1
+ * @version:    0.5
  * @changelog:  0.1 initial release
- * 
+ *              0.2 new pattern with webdriverio
+ *              0.5 new pattern with puppeteer
+ *
  */
-class Likemode_realistic extends Manager_state{
+const Manager_state = require("../common/state").Manager_state;
+class Likemode_classic extends Manager_state {
     constructor(bot, config, utils) {
         super();
         this.bot = bot;
         this.config = config;
         this.utils = utils;
-
+        this.LOG_NAME = "like";
+        this.Manager_state = require("../common/state").Manager_state;
+        this.STATE = require("../common/state").STATE;
+        this.STATE_EVENTS = require("../common/state").EVENTS;
+        this.Log = require("../logger/Log");
         this.cache_hash_tags = [];
-        this.log = new Log(LOG_NAME);
+        this.log = new this.Log(this.LOG_NAME);
     }
 
     /**
      * Get photo url from cache
      * @return {string} url
      */
-    get_photo_url(){
+    get_photo_url() {
         let photo_url = "";
         do {
             photo_url = this.cache_hash_tags.pop();
         } while ((typeof photo_url === "undefined" || photo_url.indexOf("tagged") === -1) && this.cache_hash_tags.length > 0);
+
         return photo_url;
     }
 
     /**
-     * likemode_realistic: Open Hashtag
+     * likemode_classic: Open Hashtag
      * =====================
      * Get random hashtag from array and open page
      *
      */
     async like_open_hashtagpage() {
         let hashtag_tag = this.utils.get_random_hash_tag();
-        this.log.info(`current hashtag ${hashtag_tag}`);
+        this.log.info(`current hashtag  ${hashtag_tag}`);
+
         try {
             await this.bot.goto("https://www.instagram.com/explore/tags/" + hashtag_tag + "/");
         } catch (err) {
@@ -58,11 +58,11 @@ class Likemode_realistic extends Manager_state{
 
         this.utils.sleep(this.utils.random_interval(4, 8));
 
-        await this.utils.screenshot(LOG_NAME, "last_hashtag");
+        await this.utils.screenshot(this.LOG_NAME, "last_hashtag");
     }
 
     /**
-     * likemode_realistic: Open Photo
+     * likemode_classic: Open Photo
      * =====================
      * Open url of photo and cache urls from hashtag page in array
      *
@@ -95,7 +95,7 @@ class Likemode_realistic extends Manager_state{
             } catch (err) {
                 this.cache_hash_tags = [];
                 this.log.error(`like_get_urlpic error ${err}`);
-                await this.utils.screenshot(LOG_NAME, "like_get_urlpic_error");
+                await this.utils.screenshot(this.LOG_NAME, "like_get_urlpic_error");
             }
         } else {
             photo_url = this.get_photo_url();
@@ -113,7 +113,7 @@ class Likemode_realistic extends Manager_state{
     }
 
     /**
-     * likemode_realistic: Love me
+     * likemode_classic: Love me
      * =====================
      * Click on heart and verify if instagram not (soft) ban you
      *
@@ -125,17 +125,18 @@ class Likemode_realistic extends Manager_state{
             await this.bot.waitForSelector("main article:nth-child(1) section:nth-child(1) a:nth-child(1)");
             let button = await this.bot.$("main article:nth-child(1) section:nth-child(1) a:nth-child(1)");
             await button.click();
-            this.log.warning("<3");
+            this.log.info("<3");
+            this.emit(this.STATE_EVENTS.CHANGE_STATUS, this.STATE.OK);
         } catch (err) {
             if (this.utils.is_debug())
                 this.log.debug(err);
-            
-           this.log.warning("</3");
+
+            this.log.warning("</3");
+            this.emit(this.STATE_EVENTS.CHANGE_STATUS, this.STATE.ERROR);
         }
 
         this.utils.sleep(this.utils.random_interval(4, 8));
-
-        await this.utils.screenshot(LOG_NAME, "last_like_after");
+        await this.utils.screenshot(this.LOG_NAME, "last_like_after");
     }
 
     /**
@@ -144,44 +145,45 @@ class Likemode_realistic extends Manager_state{
      *
      */
     async start() {
-        this.log.info("realistic");
+        this.log.info("classic");
 
         let today = "";
+        let t1, t2, sec, sec_min, sec_max;
+        sec_min = parseInt(86400 / this.config.bot_likeday_max);
+        sec_max = parseInt(86400 / this.config.bot_likeday_min);
 
         do {
             today = new Date();
-            this.log.info("time night: " + (parseInt(today.getHours() + "" + (today.getMinutes() < 10 ? "0" : "") + today.getMinutes())));
-            if (parseInt(today.getHours() + "" + (today.getMinutes() < 10 ? "0" : "") + today.getMinutes()) >= (this.config.bot_sleep_night).replace(":", "")) {
+            t1 = new Date(today.getFullYear(), today.getMonth(), today.getDate(), today.getHours(), today.getMinutes(), today.getSeconds());
 
-                this.log.info("loading... " + new Date(today.getFullYear(), today.getMonth(), today.getDate(), today.getHours(), today.getMinutes(), today.getSeconds()));
-                this.log.info( "cache array size " + this.cache_hash_tags.length);
+            this.log.info("loading... " + new Date(today.getFullYear(), today.getMonth(), today.getDate(), today.getHours(), today.getMinutes(), today.getSeconds()));
+            this.log.info("cache array size " + this.cache_hash_tags.length);
+            if (this.cache_hash_tags.length <= 0)
+                await this.like_open_hashtagpage();
 
-                if (this.cache_hash_tags.length <= 0)
-                    await this.like_open_hashtagpage();
+            this.utils.sleep(this.utils.random_interval(4, 8));
 
-                this.utils.sleep(this.utils.random_interval(4, 8));
+            await this.like_get_urlpic();
+            this.utils.sleep(this.utils.random_interval(4, 8));
 
-                await this.like_get_urlpic();
+            await this.like_click_heart();
 
-                this.utils.sleep(this.utils.random_interval(4, 8));
+            if (this.cache_hash_tags.length < 9) //remove popular photos
+                this.cache_hash_tags = [];
 
-                await this.like_click_heart();
+            today = new Date();
+            t2 = new Date(today.getFullYear(), today.getMonth(), today.getDate(), today.getHours(), today.getMinutes(), today.getSeconds());
+            sec = Math.abs((t1.getTime() - t2.getTime()) / 1000);
 
-                if (this.cache_hash_tags.length < 9 || this.is_ready()) //remove popular photos
-                    this.cache_hash_tags = [];
-
-                if (this.cache_hash_tags.length <= 0 && this.is_not_ready()) {
-                    this.log.info("finish fast like, bot sleep " + this.config.bot_fastlike_min + "-" + this.config.bot_fastlike_max + " minutes");
-                    this.cache_hash_tags = [];
-                    this.utils.sleep(this.utils.random_interval(60 * this.config.bot_fastlike_min, 60 * this.config.bot_fastlike_max));
-                }
+            if (sec < sec_min && this.get_status() === 1) {
+                this.log.info("seconds of loop " + sec + "... for miss ban bot wait " + (sec_min - sec) + "-" + (sec_max - sec));
+                this.utils.sleep(this.utils.random_interval(sec_min - sec, sec_max - sec));
             } else {
-                this.log.info("is night, bot sleep");
-                this.utils.sleep(this.utils.random_interval(60 * 4, 60 * 5));
+                this.cache_hash_tags = [];
             }
         } while (true);
     }
 
 }
 
-module.exports = (bot, config, utils) => { return new Likemode_realistic(bot, config, utils); };
+module.exports = (bot, config, utils) => { return new Likemode_classic(bot, config, utils); };
