+++ modules/mode/likemode_realistic.js
@@ -1,39 +1,34 @@
-const LOG_NAME = "like";
-
-const Manager_state = require("../modules/base/state").Manager_state;
-const STATE = require("../modules/base/state").STATE;
-const STATE_EVENTS = require("../modules/base/state").EVENTS;
-
-// log
-const Log = require("./logger/Log");
-
 /**
- * MODE: fdfmode_classic
+ * MODE: likemode_realistic
  * =====================
- * Defollow all your following (not defollow users in whitelist) | 90 defollow/hour.
+ * Select random hashtag from config list, like fast 10-12 photo and sleep 15-20min. Sleep at night | 400-600 like/day.
  *
  * @author:     Patryk Rzucidlo [@ptkdev] <support@ptkdev.io> (https://ptkdev.it)
  * @license:    This code and contributions have 'GNU General Public License v3'
  * @version:    0.1
  * @changelog:  0.1 initial release
- *
+ * 
  */
-class Fdfmode_classic extends Manager_state{
+const Manager_state = require("../common/state").Manager_state;
+class Likemode_realistic extends Manager_state {
     constructor(bot, config, utils) {
         super();
         this.bot = bot;
         this.config = config;
         this.utils = utils;
-
         this.cache_hash_tags = [];
-        this.log = new Log(LOG_NAME);
+        this.LOG_NAME = "like_realistic";
+        this.STATE = require("../common/state").STATE;
+        this.STATE_EVENTS = require("../common/state").EVENTS;
+        this.Log = require("../logger/Log");
+        this.log = new this.Log(this.LOG_NAME);
     }
 
     /**
      * Get photo url from cache
      * @return {string} url
      */
-    get_photo_url(){
+    get_photo_url() {
         let photo_url = "";
         do {
             photo_url = this.cache_hash_tags.pop();
@@ -42,7 +37,7 @@ class Fdfmode_classic extends Manager_state{
     }
 
     /**
-     * fdfmode_classic: Open Hashtag
+     * likemode_realistic: Open Hashtag
      * =====================
      * Get random hashtag from array and open page
      *
@@ -58,11 +53,11 @@ class Fdfmode_classic extends Manager_state{
 
         this.utils.sleep(this.utils.random_interval(4, 8));
 
-        await this.utils.screenshot(LOG_NAME, "last_hashtag");
+        await this.utils.screenshot(this.LOG_NAME, "last_hashtag");
     }
 
     /**
-     * fdfmode_classic: Open Photo
+     * likemode_realistic: Open Photo
      * =====================
      * Open url of photo and cache urls from hashtag page in array
      *
@@ -95,11 +90,12 @@ class Fdfmode_classic extends Manager_state{
             } catch (err) {
                 this.cache_hash_tags = [];
                 this.log.error(`like_get_urlpic error ${err}`);
-                await this.utils.screenshot(LOG_NAME, "like_get_urlpic_error");
+                await this.utils.screenshot(this.LOG_NAME, "like_get_urlpic_error");
             }
         } else {
             photo_url = this.get_photo_url();
-            this.log.info(`current photo url from cache  ${photo_url}`);
+
+            this.log.info(`current photo url from cache ${photo_url}`);
             this.utils.sleep(this.utils.random_interval(4, 8));
 
             try {
@@ -108,12 +104,11 @@ class Fdfmode_classic extends Manager_state{
                 this.log.error(`goto ${err}`);
             }
         }
-
         this.utils.sleep(this.utils.random_interval(4, 8));
     }
 
     /**
-     * fdfmode_classic: Love me
+     * likemode_realistic: Love me
      * =====================
      * Click on heart and verify if instagram not (soft) ban you
      *
@@ -121,23 +116,23 @@ class Fdfmode_classic extends Manager_state{
     async like_click_heart() {
         this.log.info("try heart like");
 
-        let selector_heart = "main article:nth-child(1) section:nth-child(1) a:nth-child(1)";
         try {
-            await this.bot.waitForSelector(selector_heart);
-            let button = await this.bot.$(selector_heart);
+            await this.bot.waitForSelector("main article:nth-child(1) section:nth-child(1) a:nth-child(1)");
+            let button = await this.bot.$("main article:nth-child(1) section:nth-child(1) a:nth-child(1)");
             await button.click();
             this.log.info("<3");
+            this.emit(this.STATE_EVENTS.CHANGE_STATUS, this.STATE.OK);
         } catch (err) {
             if (this.utils.is_debug())
                 this.log.debug(err);
 
-            this.log.info("</3");
-            this.emit(STATE_EVENTS.CHANGE_STATUS, STATE.ERROR);
+            this.log.warning("</3");
+            this.emit(this.STATE_EVENTS.CHANGE_STATUS, this.STATE.ERROR);
         }
 
         this.utils.sleep(this.utils.random_interval(4, 8));
 
-        await this.utils.screenshot(LOG_NAME, "last_like_after");
+        await this.utils.screenshot(this.LOG_NAME, "last_like_after");
     }
 
     /**
@@ -148,10 +143,11 @@ class Fdfmode_classic extends Manager_state{
     async start() {
         this.log.info("realistic");
 
-        do {
-            let today = new Date();
-            this.log.info("time night: " + (parseInt(today.getHours() + "" + (today.getMinutes() < 10 ? "0" : "") + today.getMinutes())));
+        let today = "";
 
+        do {
+            today = new Date();
+            this.log.info("time night: " + (parseInt(today.getHours() + "" + (today.getMinutes() < 10 ? "0" : "") + today.getMinutes())));
             if (parseInt(today.getHours() + "" + (today.getMinutes() < 10 ? "0" : "") + today.getMinutes()) >= (this.config.bot_sleep_night).replace(":", "")) {
 
                 this.log.info("loading... " + new Date(today.getFullYear(), today.getMonth(), today.getDate(), today.getHours(), today.getMinutes(), today.getSeconds()));
@@ -185,4 +181,4 @@ class Fdfmode_classic extends Manager_state{
 
 }
 
-module.exports = (bot, config, utils) => { return new Fdfmode_classic(bot, config, utils); };
+module.exports = (bot, config, utils) => { return new Likemode_realistic(bot, config, utils); };
